+++ includes/content-registration/custom-post-types-registration.php
@@ -388,7 +388,9 @@ function generate_custom_post_type_args( array $args ): array {
 		'rest_controller_class' => __NAMESPACE__ . '\REST_Posts_Controller',
 		'rewrite'               => [
 			'with_front' => $args['with_front'] ?? true,
+			'slug'       => $args['slug'] ?? sanitize_key( $args['singular'] ),
 		],
+		'has_archive'           => (bool) ( $args['has_archive'] ?? false ),
 	];
 
 	if ( ! empty( $args['api_visibility'] ) && 'private' === $args['api_visibility'] ) {
@@ -587,13 +589,12 @@ function register_content_fields_with_graphql( TypeRegistry $type_registry ) {
 						if ( $is_repeatable_field ) {
 							return array_map(
 								function( $media_id ) use ( $context ) {
-									return DataSource::resolve_post_object( (int) $media_id, $context );
+									return $context->get_loader( 'post' )->load_deferred( (int) $media_id );
 								},
 								$value
 							);
 						}
-
-						return DataSource::resolve_post_object( (int) $value, $context );
+						return $context->get_loader( 'post' )->load_deferred( (int) $value );
 					}
 
 					// If the multiple choice field has no saved data, return an empty array.