+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Atlas Content Modeler Changelog
 
+## Unreleased
+### Added
+- Added a “Post Type Archives” option on the edit model screen to control custom post type archives (The WordPress `has_archive` setting. false by default).
+
 ## 0.22.0 - 2022-09-22
 ### Added
 - New `wp acm model change-id` WP-CLI command to change a model's ID and migrate existing posts to the new ID.